requirements.txt
django==4.0.6
djangorestframework==3.14.0
django-cors-headers==3.14.0
gunicorn==20.1.0
python-dotenv==1.0.0
request=2.28.1
# whitenoise==5.3.0
# psycopg2-binary==2.9.5
# dj-database-url==1.0.0
